add ChairsStore, Simulation, Provider to RoomOverview, refactor Sideboard
