+++ src/components/Simulation/Simulation.scss
@@ -1,0 +1,3 @@
+.Simulation {
+    flex: 1;
+}
