--- src/stores/Chairs.js
@@ -1,3 +1,0 @@
-const chairs = [];
-
-export default chairs;
